+++ NEWS.md
@@ -8,6 +8,7 @@
   * Check margin notes & captions finish with a full stop
   * Add the course package version to the final page
   * Remove old files in final/ when make final
+  * Check for correct texlive version
 
 # 0.5.0
   * Extract latex tags using xparse from latex